Fixing wholesale price datatype

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/data/model/params/add/Wholesale.kt
@@ -12,7 +12,7 @@ data class Wholesale(
         var minQty: Int = 0,
         @SerializedName("price")
     @Expose
-        var price: Float = 0F
+        var price: Double = 0.toDouble()
 ) : Parcelable
 
 @Parcelize
